Add the new compiler syntax and split build/host
As this contains compiled code, add the new `compiler` Jinja syntax and
note that it is C code. This will help in migration to the new
compilers. Also split `build` and `host` so that `build` contains only
build tools like the compilers and `host` contains only things that will
show up in the run time environment.